requirements.txt
streamlit
pdfminer.six
sentence-transformers
scikit-learn
openai
python-dotenv
pypdf2